CURLHandle-class {RCurl}	R Documentation

Class "CURLHandle" for synchronous HTTP requests

Description

This is the basic class used for performing HTTP requests in the RCurl package. In R, this is a reference to a C-level data structure so we treat it as an opaque data type. However, essentially we can think of this as an with a set of options that persists across calls, i.e. HTTP requests. The numerous options that one can set can be see via getCurlOptionsConstants. The object can keep a connection to a Web server open for a period of time across calls.

This class differs from MultiCURLHandle-class as it can handle only one request at a time and blocks until the request is completed (normally or abnormally). The other class can handle asynchronous, multiple connections.

Objects from the Class

A virtual Class: No objects may be created from it.

Extends

Class "oldClass", directly.

Author(s)

Duncan Temple Lang

References

http://curl.haxx.se, the libcurl web site.

See Also

getURL, getForm, postForm dupCurlHandle, getCurlHandle, MultiCURLHandle-class

[Package RCurl version 1.95-4.11 Index]