clone {RCurl}	R Documentation

Clone/duplicate an object

Description

This is a generic function and methods for making a copy of an object such as a curl handle, C-level pointer to a file, etc.

Usage

clone(x, ...)

Arguments

x
the object to be cloned.

...
additional parameters for methods

Value

Typically, an object of the same class and “value” as the input - x.

Author(s)

Duncan Temple Lang

See Also

dupCurlHandle

Examples

 h = getCurlHandle(verbose = TRUE)
 other = dupCurlHandle(h)
 curlSetOpt(curl = h, verbose = FALSE)

[Package RCurl version 1.95-4.11 Index]