curlVersion {RCurl}	R Documentation

Information describing the Curl library

Description

This function queries the Curl library to provide information about its characteristics when it was compiled. This tells the user about its capabilities and can be used to determine strategies.

Usage

curlVersion(id = 0)

Arguments

id
an integer value between 0 and 3 inclusive. The idea is that one specifies the identifier for the version of interest. In fact, all values seem to yield the same result.

Value

A list

age
integer giving the number of this libcurl, 0 is FIRST, 1 is SECOND, 2 is THIRD

version
the version identifier as a string, e.g. "7.12.0"

version_num
the value as an integer

host
the machine on which the libcurl was configured/built.

features
a named integer vector of bits indicating what features of libcurl were configured and built into this version. These are features such as ipv6, ssl, libz, largefile, ntlm (Microsoft "authorization").

ssl_version
the string identifying the SSL version.

ssl_version_num
the number identifying the SSL version

libz_version
the string identifying the version of libz.

protocols
a character vector of the supported HTTP protocols, e.g. http, https, ftp, ldap, gopher, telnet

ares
name of the asynchronous DNS (domain name service) lookup library. This is often simply the empty string indicating it is not there.

ares_num
the number for the ares library

libidn
the name of the IDN (internationalized domain names) library being used. This field only appears in version 3 of libcurl. If you are using version 2 (e.g. curl-7.11.2), this will be NA. An empty string indicates that the field is present, but has no value.

See the man page for curl_version_info for a description of these fields. features in R is a named integer vector detailing the different features.

Author(s)

Duncan Temple Lang <duncan@wald.ucdavis.edu>

References

Curl homepage http://curl.haxx.se

See Also

curl_version_info in the libcurl documentation.

Examples

  curlVersion()

[Package RCurl version 1.95-4.11 Index]