getCurlInfo {RCurl}	R Documentation

Access information about a CURL request

Description

This function provides access to data about a previously executed CURL request that is accessible via a CURLHandle object. This means, of course, that one must have access to the CURLHandle object. The information one can get includes items such as the name of the file (potentially containing redirects), download time,

See getCurlInfoConstants for the names of the possible fields.

Usage

getCurlInfo(curl, which = getCurlInfoConstants())
getCurlInfoConstants()

Arguments

curl
the CURLHandle object used to perform the request. This is a reference to an opaque internal C-level data structure that is provided and used by libcurl to make a request.

which
identifiers for the elements of interest. These can be specified by integer value or by name. The names are matched against those in the getCurlInfoConstants

.

Details

This is an interface to the get_curl_info routine in the libcurl package.

Value

A named list whose elements correspond to the requested fields. The names are expanded to match the names of these fields and the values are either strings or integer values.

Author(s)

Duncan Temple Lang <duncan@wald.ucdavis.edu>

References

Curl homepage http://curl.haxx.se

See Also

curlPerform getURL getCurlHandle

Examples

 if(url.exists("http://www.omegahat.net/RCurl/index.html")) {
    curl = getCurlHandle()
    getURL("http://www.omegahat.net/RCurl/index.html", curl = curl)
    getCurlInfo(curl)
    rm(curl)  # release the curl!
 }

[Package RCurl version 1.95-4.11 Index]