url.exists {RCurl}	R Documentation

Check if URL exists

Description

This functions is analogous to file.exists and determines whether a request for a specific URL responds without error. We make the request but ask the server not to return the body. We just process the header.

Usage

url.exists(url, ..., .opts = list(...),
            curl = getCurlHandle(.opts = .opts),
             .header = FALSE)

Arguments

url
the URL whose existence we are to test

...
name = value pairs of Curl options.

.opts
a list of name = value pairs of Curl options.

curl
a Curl handle that the caller can specify if she wants to reuse an existing handle, e.g. with different options already specified or that has previously established a connection to the Web server

.header
a logical value that if TRUE causes the header information to be returned.

Details

This makes an HTTP request but with the nobody option set to FALSE so that we don't actually retrieve the contents of the URL.

Value

If .header is FALSE, this returns TRUE or FALSE indicating whether the request was successful (had a status with a value in the 200 range).

If .header is TRUE

Author(s)

Duncan Temple Lang

References

HTTP specification

See Also

curlPerform

Examples

 url.exists("http://www.omegahat.net/RCurl")
 try(url.exists("http://www.omegahat.net/RCurl-xxx"))

[Package RCurl version 1.95-4.11 Index]