AUTH_ANY {RCurl}	R Documentation

Constants for identifying Authentication Schemes

Description

These variables are symbolic constants that allow use to specify different combinations of schemes for HTTP authentication in a request to a Web server. We can combine them via the | operator to indicate that libcurl should try them in order until one works.

Examples

   AUTH_BASIC | AUTH_DIGEST

[Package RCurl version 1.95-4.11 Index]