CurlFeatureBits {RCurl}	R Documentation

Constants for libcurl

Description

These are enums and bit fields defining constants used in libcurl and used in R to specify values symbolically.

Usage

CurlFeatureBits

Format

named integer vectors. The names give the symbolic constants that can be used in R and C code. These are mapped to their integer equivalents and used in C-level computations.

Source

libcurl (see http://curl.haxx.se)

[Package RCurl version 1.95-4.11 Index]