CURLEnums {RCurl}	R Documentation

Classes and coercion methods for enumerations in libcurl

Description

These are classes and coercion methods for enumeration types in RCurl corresponding to symbolic constants in libcurl. The actual constants are not exported, but are defined within the package. So we can use them with code such as RCurl:::CURLINFO_DATA_IN.

Author(s)

Duncan Temple Lang

[Package RCurl version 1.95-4.11 Index]