getFormParams {RCurl}	R Documentation

Extract parameters from a form query string

Description

This function facilitates getting the parameter names and values from a URL that is an parameterized HTML query.

This is motivated by a function from Chris Davis and Delft University.

Usage

getFormParams(query, isURL = grepl("^(http|\\?)", query))

Arguments

query
the query string or full URL containing the query

isURL
a logical value. If TRUE, query is the full URL and we need to extract the sub-string representing the parameters. If isURL is FALSE, then query is assumed to be just the string containing the parameters.

Value

A named character vector giving the parameter values The names are the parameter names.

Author(s)

Duncan Temple Lang

Examples

if(url.exists("http://www.omegahat.net/foo/bob.R")) {

  getFormParams("http://www.omegahat.net/foo/bob.R?xyz=1&abc=verylong")

  getFormParams("xyz=1&abc=verylong")
  getFormParams("xyz=1&abc=&on=true")
  getFormParams("xyz=1&abc=")
}

[Package RCurl version 1.95-4.11 Index]